Check for Int

+++ source/fast_pack.d
@@ -251,10 +251,10 @@ public:
     pragma(inline, true)
     RectangleType getRectangle(RectangleTypeIntegral)(immutable T key) {
         // This allows you to automatically downcast and insert into custom types.
-        static assert(is(typeof(RectangleTypeIntegral.x) == float), "x must be floating point.");
-        static assert(is(typeof(RectangleTypeIntegral.y) == float), "y must be floating point.");
-        static assert(is(typeof(RectangleTypeIntegral.w) == float), "w must be floating point.");
-        static assert(is(typeof(RectangleTypeIntegral.h) == float), "h must be floating point.");
+        static assert(is(typeof(RectangleTypeIntegral.x) == int), "x must be floating point.");
+        static assert(is(typeof(RectangleTypeIntegral.y) == int), "y must be floating point.");
+        static assert(is(typeof(RectangleTypeIntegral.w) == int), "w must be floating point.");
+        static assert(is(typeof(RectangleTypeIntegral.h) == int), "h must be floating point.");
 
         const(FloatingRectangle)* thisRectangle = key in floatingLookupTable;
 
@@ -289,8 +289,8 @@ public:
     /// Returns: TexturePoints!Vec2Type
     TexturePoints!Vec2Type getTexturePoints(Vec2TypeIntegral)(immutable T key) {
         // This allows you to automatically downcast and insert into custom types.
-        static assert(is(typeof(Vec2TypeIntegral.x) == float), "x must be floating point.");
-        static assert(is(typeof(Vec2TypeIntegral.y) == float), "y must be floating point.");
+        static assert(is(typeof(Vec2TypeIntegral.x) == int), "x must be floating point.");
+        static assert(is(typeof(Vec2TypeIntegral.y) == int), "y must be floating point.");
 
         const(FloatingRectangle)* thisRectangle = key in floatingLookupTable;
